\name{GlynnIchino}
\alias{GlynnIchino}
\docType{data}
\title{Glynn and Ichino (2014) Replication Data}

\description{Replication data for Glynn and Ichino (2014). Data structure is panels of Latin American countries from 1983-2007.}

\usage{GlynnIchino}


\format{A dataframe with the following variables: 
	\itemize{
		\item{\code{country}: country identifier (string).}
		\item{\code{set}: set identifier (numeric).}
		\item{\code{treat}: binary treatment indicator, where treatment denotes (insert here).}
		\item{\code{rank}: within-set unit rank.}
		}
}
		
	
\source{
cite paper here.
}

\seealso{\code{\link{quade}}. Run \code{demo(GlynnIchino)} for examples which use \code{GlynnIchino}.}
\keyword{causal inference}
\keyword{datasets}
